feat: unix is kinda working

+++ remotefs-fuse/src/driver/unix/inode.rs
@@ -5,20 +5,28 @@ pub type Inode = u64;
 
 type Database = HashMap<Inode, PathBuf>;
 
+const ROOT_INODE: Inode = 1;
+
 /// A database to map inodes to files
 ///
 /// The database is saved to a file when the instance is dropped
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Clone)]
 pub struct InodeDb {
     database: Database,
 }
 
 impl InodeDb {
     /// Load [`InodeDb`] from a file
+    ///
+    /// It will initialize an empty database with only one inode set: the root inode which has always the value 1
     pub fn load() -> Self {
-        Self {
+        let mut db = Self {
             database: Database::new(),
-        }
+        };
+
+        db.put(ROOT_INODE, PathBuf::from("/"));
+
+        db
     }
 
     /// Check if the database contains an inode
@@ -28,11 +36,17 @@ impl InodeDb {
 
     /// Put a new inode into the database
     pub fn put(&mut self, inode: Inode, path: PathBuf) {
+        debug!("inode {inode} -> {}", path.display());
         self.database.insert(inode, path);
     }
 
     /// Forget an inode
     pub fn forget(&mut self, inode: Inode) {
+        if inode == ROOT_INODE {
+            error!("tried to roget 1");
+            return;
+        }
+
         self.database.remove(&inode);
     }
 
@@ -50,14 +64,26 @@ mod test {
 
     #[test]
     fn test_inode_db() {
-        let mut db = InodeDb::default();
+        let mut db = InodeDb::load();
 
-        db.put(1, PathBuf::from("/test"));
-        assert_eq!(db.get(1), Some(Path::new("/test")));
-        assert_eq!(db.has(1), true);
+        // should have root inode
+        assert_eq!(db.has(ROOT_INODE), true);
+        assert_eq!(db.get(ROOT_INODE), Some(Path::new("/")));
 
-        db.forget(1);
-        assert_eq!(db.get(1), None);
-        assert_eq!(db.has(1), false);
+        db.put(3, PathBuf::from("/test"));
+        assert_eq!(db.get(3), Some(Path::new("/test")));
+        assert_eq!(db.has(3), true);
+
+        db.forget(3);
+        assert_eq!(db.get(3), None);
+        assert_eq!(db.has(3), false);
+    }
+
+    #[test]
+    fn test_should_not_forget_root() {
+        let mut db = InodeDb::load();
+
+        db.forget(ROOT_INODE);
+        assert_eq!(db.has(ROOT_INODE), true);
     }
 }
